+++ packages/modules/create-imodel/src/components/create-imodel/CreateIModel.test.tsx
@@ -2,11 +2,23 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { toaster } from "@itwin/itwinui-react";
+import { useToaster } from "@itwin/itwinui-react";
 import { act, fireEvent, render } from "@testing-library/react";
 import React from "react";
 
 import { CreateIModel } from "./CreateIModel";
+
+jest.mock("@itwin/itwinui-react", () => {
+  const actual = jest.requireActual("@itwin/itwinui-react");
+
+  return {
+    ...actual,
+    useToaster: jest.fn().mockReturnValue({
+      positive: jest.fn(),
+      negative: jest.fn(),
+    }),
+  };
+});
 
 describe("CreateIModel", () => {
   const mockedimodel = { iModel: { id: "dd", name: "name" } };
@@ -28,7 +40,7 @@ describe("CreateIModel", () => {
 
   it("should create an iModel", async () => {
     const successMock = jest.fn();
-    toaster.positive = jest.fn();
+    const toaster = useToaster();
 
     const { getByText, container } = render(
       <CreateIModel
@@ -76,7 +88,7 @@ describe("CreateIModel", () => {
     const errorMock = jest.fn();
     const error = new Error("Fail");
     fetchMock.mockImplementationOnce(() => Promise.reject(error));
-    toaster.negative = jest.fn();
+    const toaster = useToaster();
 
     const { getByText, container } = render(
       <CreateIModel
@@ -130,7 +142,7 @@ describe("CreateIModel", () => {
     const errorMock = jest.fn();
     const error = { error: { code: "iModelExists" } };
     fetchMock.mockImplementationOnce(() => Promise.reject(error));
-    toaster.negative = jest.fn();
+    const toaster = useToaster();
 
     const { getByText, container } = render(
       <CreateIModel